feat: refactor input fields
